Moved cleanup to SharedObjects.DestroyObjects().
This is a better place to remove the hooks, because
the destructor (where they were) cannot fire off
while the hook itself is preventing orphaning.

I left the old location in place as well, because
I think it can still fire too and it works
as a fallback if we don't always remember to remove
those hooks.