Fixes #2086 onradiochange bug
The problem was that the test to see whether or not OnRadioChange should be called was buried inside the test to see whether or not OnToggle should be called. If the conditions weren't right to cause OnToggle, then it also would never check for OnRadioChange.

It just so happened that in my tests I had been trying to test every callback - so I had them all enabled to see which fires when, which meant it never tested the case where OnRadioChange was defined but OnToggle was not.