Fixes #2082 by allowing NoDelegate w/out Shared
Fixes #2082.

The problem was that the GUI code was trying to make use of
a feature of the ``NoDelegate`` type which it turns out
never really worked right, ``NoDelegate.Instance``.

NoDelegate.Instance can't refer to which Cpu it's part of (to
do so it would have to have one copy per ``SharedObjects``,
rather than one gobally static instance.)  So it tries to
construct a ``NoDelegate`` with a null ``Cpu``, under the
premise that you don't need a real ``Cpu`` defined if it
can never be called.  But it turned out
that never really worked because the NoDelegate constructor,
when making the very first ``instance``, still tries to call
``Cpu.GetCurrentContext()`` (on the null ``Cpu``) as part
of passing off to the ``UserDelegate`` base constructor, here:

```
        public NoDelegate(ICpu cpu) : base(cpu, cpu.GetCurrentContext(), -1, false)
        {
        }
```

Fixing this meant making it so that NoDelegate is actually able
to tolerate being constructed with a null ``Cpu``, and that meant
having to insert a few extra safety checks here and there.