Update kOS.netkan
"x_netkan_epoch" gives us a numbering restart allowing us to jump over the 1.3.3.1 release. The epoch is hidden by default in the CKAN client.
"x_netkan_override" forces the min ksp version for the 1.1.3.1 release, which is still "1.0.0" in the download

I'll duplicate the 1.1.3.0 .ckan to 1:1.1.3.0 in the CKAN repo and change any others with the 1.0.0 ksp_min version.