Fixes #2297 - misplaced Jump in trigger locks.
It turns out there was a problem buried deep in
the compiler, and it was introduced first with commit
number #9aca5d8, with Pull Request #2156.

That commit altered the boilerplate code inserted for
triggered locks (steering, throttle, etc), reducing
the number of instructions that existed in the
boilerplate code chunk.  But just prior to those
edits, on the line above, there was a br.true +4.
Becuse of those edits, there was now one fewer
opcode to skip over and the br.true +4 needed to be
changed to a br.true +3, and it hadn't been.

The result is that when the function $steering*
already exists, and it tries to skip over the
setting of it, it skips the wrong distance missing
a "push", and thus everything on the stack is
mis-aligned after that.

The reason it hasn't been noticed before is that usually,
this case doesn't happen in typical cases.  It requires
running a boot file that runs a compiled KSM file to
encounter the case where that br.true will fire off. Most
of the time it was false and not executing, thus the wrong
size of how far it jumps wasn't having any effect.