fix possible race condition preventing detection
The ability to detect the need to re-default the lock steering
doesn't seem to be always happening right when the part count
changes.  Sometimes the problem isn't detected till a moment later,
which I assume must be a race condition.  So I added the ability
to flag the problem when detected to force it to go through
the work again to check for stale controls.  This seems to fix
the problem more consistently.