Fixes #2443: ISCONNECTED() shouldn't crash on debris.
The stock CommNet problem, that things classified as debris
are missing properties (they are null) on scene load isn't
something we can fix.  It's probably the stock game's way
of reducing system load by culling out all debris from the
comms system when loading a scene.  The weird fact that even
changing the vessel type doesn't really do anything until
the next scene reload isn't something we can do anything about
either.

But what we can do it make sure that calling ISCONNECTED()
merely returns false, rather than crashing with nullref, when
the vessel is not a participant in the stock CommNet system.

As I went to mention this in the docs, I also noticed some
out of date commentary in the docs. It claimed that ISCONNECTED
is always true in stock, and can only ever be false if you use
Remote Tech.  This needed fixing because it was written before
stock CommNet existed.  So I went through that section and
re-wrote those paragraphs to ensure the 3 options are expressly
described one at a time in bullet points (stock no comms, stock
with CommNet, and RemoteTech).