Fixes #2449 by slightly altering input locking.
The bug was caused by this:

TermWindow WAS unlocking its own inputlock flags when OnDestroy()
was being called, which seems like it should avoid the problem.
However TermWindow is a derived class from KOSManagedWindow, which
has its *own* input locks, and that base class's LoseFocus was being
skipped in the OnDestroy() hook of TermWindow.  So TermWindow's locks
were being cleared, but the base KOSManagedWindow's locks were not.