Fixes #2448 - refactor serialization so geocoords work
Fixes #2448:
This refactor is described better by issue #2448's comments, where
I wrote down a detailed list of what I was doing before I did it.

In a nutshell, previous to this refactor, classes implementing
IDumper had a secret (unenforced) requirement to always have
a default constructor or the serialiazation would fail to
create the object on reading the file/message.  That was
because the algorithm used in reading was:
  1 - create instance with default constructor.
  2 - populate instance using LoadDump() call.
Because interfaces cannot have static things in them, we
couldn't do it properly by saying IDumper is a contract
to create a single factory method that does both (1) and (2)
at once.

This refactor moves the logic to using a new static method
called CreateFromDump() that does do (1) and (2) together,
and the enforcement to confirm that you do indeed have
such a method is done via reflection at runtime since
the compiler cannot do it (C# is one of the almost-OOP
languages that can't handle inheriting static things.)

Once that refactor was done, then Geocoordinates started
working, solving #2448.