Fixes #2456 by moving the dialog box
Fixes #2456

The PopupDialog.SpawnPopupDialog() call from stock KSP's API
allows you to pass in some "anchor" coordinates to move
the window - but it took trial and error to work out what
those coordinates really meant.  They weren't intuitive
(see the comment in the PR for further explanation).

But in the end I finally managed to move the window.
Note that the same call is in a spot of our code used
by another dialog window, so I had to make a way to preserve
it's previous (centered) behavior, so I had to add a little
extra stuff to make it possible to pass in x,y values to
QueueDialog and remember them with the dialog box.

It's very hard to regression-test that I didn't break that
second dialog box (the "migration dialog") because it's
only triggered if you take a very old saved game and try
to migrate it to newer kOS.  It was years ago when that was
a relevant thing and it's nearly impossible to test now
without getting some old copies of KSP.  I'm happy leaving
that part untested, as I couldn't easily make it happen, and
anyone who thinks their 3-year old KSP saved game will still
work fine today is going to run into lots of other migration
issues besides just kOS.