Fixes #2461 adding LIST FONTS command.
Fixes #2461

I originally wanted to put the font list on a suffix
of something and not add more to the old LIST blarg IN FOO
command, but there was nowhere to put it that was "static"
rather than "instance".  It didn't make sense to say something
like this to get at it, having to use a constructed instance
of something to get at the suffix:

```
local Gwin is GUI().
print Gwin:fontnames.
```

When the suffix is inherently a "static" thing.  kerboscript
sort of lacks the idea of "static" members of a Structure,
so I went with the old LIST blarg IN FOO syntax to offer
this feature.