Made BodyTarget into an IKOSTargetable again.
The word "IKOSTargetable" was accidentally deleted from
the declaration header of class BodyTarget by my PR #2460.

The SET TARGET command has a check to see if the value being
set is an actual target object or just a string name that
needs to be looked up.  That check was based on seeing if
the value could be cast to IKOSTargetable or not.  When
that was accidentally removed from BodyTarget's header,
it broke that check.