Removed vessel orbits target body check.
I had inserted a check to throw an exception if you
try to SET TARGET to an illegal body, one the game
won't allow because you currently orbit it.  In retrospect,
that removes a feature people might have been using, which
is that trying to set the target to your current SOI body is
a roundabout way to unset the target (the game will refuse the
attempt and deselect the target, but the kOS script would
be fine with it and wouldn't crash).

Changing that into an exception means previously working code
could break, so I removed the check again.