Fixes #2373
While I was at it, I also edited a few lines in the
constant documentation that needed cleaning up,
since I was in there an noticed them.

There is an API call from KSP called
CelestialBody.GetFullTemperature(Vector3d position),
but as far as I can tell it seems totally broken, or
at least is in a weird reference frame I don't know about.
So I gave up on supporting that, since I don't seem
to be able to make it work (also, other KSP mods seem
to have written their own replacement for this call as
well, indicating that they found it didn't work for them
either.)