Force all input through buffer unless flushing buffer.
Fixes #2753 (I hope - I can't test it.)

Now the only input that is allowed to immediately get
sent to the interpreter is that which came FROM the
buffer flusher routine ProcessUnconsmedInput().  It
uses flags that allow its input to come through
directly, but *everything else* that types input is
now requried to queue it (so it will pass through
ProcessUnconsumedInput() a moment later).  Hypothetically
this should fix the ability for input to skip the buffer
and jump in line.

The one exception is the immediate characters related
to terminal closing or program interruption like CTRL-C.
They are allowed to jump in out of order.  And in the
case of CTRL-C, it also flushes whatever was in the
buffer so if you typed anything ahead it will no longer
get executed after a CTRL-C like it used to.