Fixes #2836 turning nodeRef into a property
The new property has a public get but a private set.

Because of coding standards on the project, the
name had to be changed to uppercase (NodeRef)
instead of lowercase (nodeRef) when it became a
property, so watch out for that when you use this.

Also be aware that the property can be null for
nodes that have not been added to the vessel's
flight plan yet, or are for a vessel that is not
currently the active vessel.  Any use of this
property should check to see if it's null.