Fix buggy math that only appears off equator.
I thought I had solved this in the previous commit but
noticed it failed in gameplay.  My math wasn't quite right.

The problem fixed in this commit only shows up when you aren't
on the equator, which is why it wasn't seen in testing on the KSC
runway.

The issue is that I changed algorithms partway through implementing
this but didn't fully edit the code to reflect the new algorithm.
What I had was a mix of the two.

Prev algorithm was "calculate new raycast start by adding
(distance plus a bit)*aimvector to the previous start
point".

New algorithim was "Calculate new raycast start by just
using the hit location directly, plus aimvector times a little".

The old algorithm required a +=, and the new one just an =.  I missed
that edit, and that mistake caused it not to work when not on
the equator (essentially, when the Z axis of world coords isn't zero,
the bug manifested.)