Culls out thruster transforms for other part variants
Fixes #2912.

The problem only exists in KSP 1.11 and up.  In KSP 1.10.x and
earlier, it was working.

The problem was caused by the introduction of RCS part variants
that came with KSP 1.11.x.  Restock happened to exacerbate the
problem because it has more part variants, but even the part
variants that come with stock are enough to trigger the problem.

The problem is that kOS's replacement for GetPotentialTorque()
was adding up the superset of all the part variants' thruster
nozzles, not just the nozzles that actually exist in the current
part variant.  You'd think that KSP would  cull out the nozzles
for other part variants when it builds the
ModuleRCS.thrusterTransforms list, but it doesn't seem to do that,
so we have to cull them ourselves when iterating over that list.