Fixes #3002 ModuleCargoPart breaks PAW.
The fix was to only apply the ModuleCargoPart if the KSP version
is high enough to be safe, using ModuleManager to do it.

Note that ModuleManager still doesn't have a KSP version check filter,
so I had to indirectly check the version by looking to see if a known part
definition exists for a part that didn't exist prior to KSP 1.11.x.
That's an ugly kludge, but it does seem to be what all the other
modders who ran into this same problem did.