Enable check that prevents masking built-in identifiers.
The eventual plan is to be able to toggle this feature off
in two ways - one is the CONFIG:CLOBBERBUILTINS setting for the
default, and the other is a compiler directive @CLOBBERBUILTINS
that can override the default on a per-file basis.  That compiler
directive isn't there yet in this PR, but you can test turning
the feature on and off via the CONFIG option.

What's needed for me to make the PR merge-ready:

[ ] Add @CLOBBERBUILTIN directive, which requires some
grammar file changes which is why I didn't tackle it yet.

[ ] Add user-facing documentation about this, which means
mentioning it in several place through the docs, and
mentioning how older scripts you find on the internet
might break because of it unless you change the setting,
etc, etc, etc.