KOSCompileException gives filename info better.
When Compler.cs throws KOSCompileExceptions, it tends to
print the filename/line/col now.

However, KOSNumberParseException, which is also thrown
from the compiler, still doesn't show the info.  That's
because it's not derived from KOSCompileException, and
gets thrown by StringValue.ToNumber(), which also gets
called from contexts that aren't the compiler, so it
can't contain the file/line/col info.  Fixing that is
a bit more complex, but what this commit does is still an
improvement over what it was doing before, so it's still
worth checking in.